% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBAPlot.R
\name{ggBAPlot}
\alias{ggBAPlot}
\title{Bland-Altman plot}
\usage{
ggBAPlot(data, names_from, values_from, group_from = NULL, unit = "")
}
\arguments{
\item{data}{Data set to use for ploting - in a long format. A data.frame.}

\item{names_from, values_from}{Character strings indicating which column contains the assay identifier and the values, resp.}

\item{unit}{SI unit of the data in \code{values_from}}
}
\value{
A ggplot
}
\description{
Draws a Bland-Altman Plot for visually analyzing the agreement between two different assays using \link[ggplot2:ggplot]{ggplot2:;ggplot()}.
}
\seealso{
\link[ggplot2:ggplot]{ggplot2::ggplot()}
}
