% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupStats.R
\name{GroupStats}
\alias{GroupStats}
\title{Group statistics}
\usage{
GroupStats(
  data,
  values_from,
  group_from,
  statfun = function(x) {
     c(mean = mean(x), se = sqrt(var(x)/length(x)), n =
    length(x))
 }
)
}
\arguments{
\item{data}{Data set to use - in a wide format. If not already a data.frame, will be converted to one.}

\item{values_from, group_from}{Character strings indicating which column contains the values and the group variable(s), resp. \var{group_from} can be a scalar or a vector.}

\item{statfun}{function to use for calculation of summary statistics. if returning more than one value, must be in a list.}
}
\value{
A \link[base:ggplot]{ggplot2::ggplot()}
}
\description{
Calculates groups statistis
}
