% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificNotation.R
\name{ScientificNotation}
\alias{ScientificNotation}
\alias{get_si}
\alias{si_to_exponent}
\alias{exponent_to_si}
\alias{get_exp_eng}
\alias{get_factor_eng}
\title{Tools for working with scientfific notations.}
\usage{
get_si(val)

si_to_exponent(letter)

exponent_to_si(exp)

get_exp_eng(val)

get_factor_eng(val)
}
\arguments{
\item{val}{A floating-point number}

\item{letter}{SI prefix or full SI string}

\item{exp}{The exponent of a number in scientific notation}
}
\value{
\code{get_si()}, \code{exponent_to_si()}: A character vector with SI prefix(es)

\code{si_to_exponent()}, \code{get_exp_eng}, \code{get_factor_eng}: A numeric vector
}
\description{
These functions perform conversions between floating-point numbers, engineering and scientific notations.
}
\section{Functions}{
\itemize{
\item \code{si_to_exponent}: Converts an SI prefix or full string into exponent

\item \code{exponent_to_si}: Converts an exponent into corresponding SI unit

\item \code{get_exp_eng}: Gets the exponent of a number in engineering notation

\item \code{get_factor_eng}: Gets the factor before the base for a number
}}

\seealso{
\link[sitools:f2si]{sitools::f2si()}
}
