% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBAPlot.R
\name{ggBAPlot}
\alias{ggBAPlot}
\title{Bland-Altman plot}
\usage{
ggBAPlot(data, names_from, values_from, unit = "")
}
\arguments{
\item{data}{Default dataset to use for plot - in a long format. If not already a data.frame, will be converted to one.}

\item{names_from, values_from}{Character strings indicating which column contains the group variable and the values, resp.}

\item{unit}{SI unit of the data in \code{values_from}}
}
\value{
A ggplot
}
\description{
Draws a Bland-Altman Plot using \link[ggplot2:ggplot]{ggplot2:;ggplot()},
}
\seealso{
\link[ggplot2:ggplot]{ggplot2:;ggplot()}
}
